quantrr:::build_sample()
