+++ pkgdown/assets/sample/index.html
@@ -2,7 +2,7 @@
 <html xmlns="http://www.w3.org/1999/xhtml" lang="en" xml:lang="en"><head>
 
 <meta charset="utf-8">
-<meta name="generator" content="quarto-1.7.33">
+<meta name="generator" content="quarto-1.7.34">
 
 <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=yes">
 
@@ -38,9 +38,9 @@ ul.task-list li input[type="checkbox"] {
 <script src="site_libs/quarto-html/tippy.umd.min.js"></script>
 <script src="site_libs/quarto-html/anchor.min.js"></script>
 <link href="site_libs/quarto-html/tippy.css" rel="stylesheet">
-<link href="site_libs/quarto-html/quarto-syntax-highlighting-ea385d0e468b0dd5ea5bf0780b1290d9.css" rel="stylesheet" class="quarto-color-scheme" id="quarto-text-highlighting-styles">
-<link href="site_libs/quarto-html/quarto-syntax-highlighting-dark-bc185b5c5bdbcb35c2eb49d8a876ef70.css" rel="stylesheet" class="quarto-color-scheme quarto-color-alternate" id="quarto-text-highlighting-styles">
-<link href="site_libs/quarto-html/quarto-syntax-highlighting-ea385d0e468b0dd5ea5bf0780b1290d9.css" rel="stylesheet" class="quarto-color-scheme-extra" id="quarto-text-highlighting-styles">
+<link href="site_libs/quarto-html/quarto-syntax-highlighting-c8ad9e5dbd60b7b70b38521ab19b7da4.css" rel="stylesheet" class="quarto-color-scheme" id="quarto-text-highlighting-styles">
+<link href="site_libs/quarto-html/quarto-syntax-highlighting-dark-befe23ebd2f54d8af2c8a89d1a1611f1.css" rel="stylesheet" class="quarto-color-scheme quarto-color-alternate" id="quarto-text-highlighting-styles">
+<link href="site_libs/quarto-html/quarto-syntax-highlighting-c8ad9e5dbd60b7b70b38521ab19b7da4.css" rel="stylesheet" class="quarto-color-scheme-extra" id="quarto-text-highlighting-styles">
 <script src="site_libs/bootstrap/bootstrap.min.js"></script>
 <link href="site_libs/bootstrap/bootstrap-icons.css" rel="stylesheet">
 <link href="site_libs/bootstrap/bootstrap-5fdbc8f45c2d0bcdc137be0dad0ef85a.min.css" rel="stylesheet" append-hash="true" class="quarto-color-scheme" id="quarto-bootstrap" data-mode="light">
@@ -346,6 +346,12 @@ window.Quarto = {
  <span class="menu-text">Widget System Risk Analysis Report</span></a>
   </div>
 </li>
+        <li class="sidebar-item">
+  <div class="sidebar-item-container"> 
+  <a href="./analysis/mfa.html" class="sidebar-item-text sidebar-link">
+ <span class="menu-text">MFA Risk Analysis Report</span></a>
+  </div>
+</li>
     </ul>
     </div>
 </nav>
@@ -359,7 +365,7 @@ window.Quarto = {
   <li><a href="#reports" id="toc-reports" class="nav-link active" data-scroll-target="#reports">Reports</a></li>
   </ul>
 </nav>
-    <h5 class="quarto-listing-category-title">Categories</h5><div class="quarto-listing-category category-default"><div class="category" data-category="">All <span class="quarto-category-count">(1)</span></div></div></div>
+    <h5 class="quarto-listing-category-title">Categories</h5><div class="quarto-listing-category category-default"><div class="category" data-category="">All <span class="quarto-category-count">(2)</span></div></div></div>
 <!-- main -->
 <main class="content" id="quarto-document-content">
 
@@ -419,7 +425,7 @@ window.Quarto = {
     </div>
 </div>
 <div class="list quarto-listing-default">
-<div class="quarto-post image-right" data-index="0" data-categories="" data-listing-date-sort="1723698000000" data-listing-file-modified-sort="1755811392942" data-listing-date-modified-sort="1755493200000" data-listing-reading-time-sort="6" data-listing-word-count-sort="1172">
+<div class="quarto-post image-right" data-index="0" data-categories="" data-listing-date-sort="1723694400000" data-listing-file-modified-sort="1756231177394" data-listing-date-modified-sort="1755489600000" data-listing-reading-time-sort="6" data-listing-word-count-sort="1172">
 <div class="thumbnail"><a href="./analysis/widgetsys.html" class="no-external">
 
 <p class="card-img-top"><img src="analysis/widgetsys_files/figure-html/risk_hist-1.png"  class="thumbnail-image card-img"/></p>
@@ -440,6 +446,31 @@ Example quantrr risk analysis of the “Widget Management System” (WMS).
 <a href="./analysis/widgetsys.html" class="no-external">
 <div class="listing-date">
 2024-08-15
+</div>
+</a>
+</div>
+</div>
+<div class="quarto-post image-right" data-index="1" data-categories="" data-listing-date-sort="1756353600000" data-listing-file-modified-sort="1757441779322" data-listing-date-modified-sort="1757390400000" data-listing-reading-time-sort="6" data-listing-word-count-sort="1131">
+<div class="thumbnail"><a href="./analysis/mfa.html" class="no-external">
+
+<p class="card-img-top"><img src="analysis/mfa_files/figure-html/risk_hist-1.png"  class="thumbnail-image card-img"/></p>
+
+</a></div>
+<div class="body">
+<h3 class="no-anchor listing-title">
+<a href="./analysis/mfa.html" class="no-external">MFA Risk Analysis Report</a>
+</h3>
+<div class="listing-categories">
+
+</div>
+<div class="delink listing-description"><a href="./analysis/mfa.html" class="no-external">
+Example quantrr risk analysis of multi-factor authentication (MFA) using real data! Originally used for the plot in my Aug 28 LinkedIn Post.
+</a></div>
+</div>
+<div class="metadata">
+<a href="./analysis/mfa.html" class="no-external">
+<div class="listing-date">
+2025-08-28
 </div>
 </a>
 </div>
@@ -869,7 +900,7 @@ Example quantrr risk analysis of the “Widget Management System” (WMS).
 <p>Copyright 2025</p>
 </div>
     <div class="nav-footer-right">
-<p>Rendered with Quarto 1.7.33</p>
+<p>Rendered with Quarto 1.7.34</p>
 </div>
   </div>
 </footer>
