+++ pkgdown/assets/sample/listings.json
@@ -2,7 +2,8 @@
   {
     "listing": "/index.html",
     "items": [
-      "/analysis/widgetsys.html"
+      "/analysis/widgetsys.html",
+      "/analysis/mfa.html"
     ]
   }
 ]
