+++ pkgdown/assets/sample/site_libs/quarto-html/quarto-syntax-highlighting-dark-befe23ebd2f54d8af2c8a89d1a1611f1.css
@@ -1,0 +1,219 @@
+/* quarto syntax highlight colors */
+:root {
+  --quarto-hl-al-color: #f07178;
+  --quarto-hl-an-color: #d4d0ab;
+  --quarto-hl-at-color: #00e0e0;
+  --quarto-hl-bn-color: #d4d0ab;
+  --quarto-hl-bu-color: #abe338;
+  --quarto-hl-ch-color: #abe338;
+  --quarto-hl-co-color: #f8f8f2;
+  --quarto-hl-cv-color: #ffd700;
+  --quarto-hl-cn-color: #ffd700;
+  --quarto-hl-cf-color: #ffa07a;
+  --quarto-hl-dt-color: #ffa07a;
+  --quarto-hl-dv-color: #d4d0ab;
+  --quarto-hl-do-color: #f8f8f2;
+  --quarto-hl-er-color: #f07178;
+  --quarto-hl-ex-color: #00e0e0;
+  --quarto-hl-fl-color: #d4d0ab;
+  --quarto-hl-fu-color: #ffa07a;
+  --quarto-hl-im-color: #abe338;
+  --quarto-hl-in-color: #d4d0ab;
+  --quarto-hl-kw-color: #ffa07a;
+  --quarto-hl-op-color: #ffa07a;
+  --quarto-hl-ot-color: #00e0e0;
+  --quarto-hl-pp-color: #dcc6e0;
+  --quarto-hl-re-color: #00e0e0;
+  --quarto-hl-sc-color: #abe338;
+  --quarto-hl-ss-color: #abe338;
+  --quarto-hl-st-color: #abe338;
+  --quarto-hl-va-color: #00e0e0;
+  --quarto-hl-vs-color: #abe338;
+  --quarto-hl-wa-color: #dcc6e0;
+}
+
+/* other quarto variables */
+:root {
+  --quarto-font-monospace: SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace;
+}
+
+/* syntax highlight based on Pandoc's rules */
+pre > code.sourceCode > span {
+  color: #f8f8f2;
+}
+
+code.sourceCode > span {
+  color: #f8f8f2;
+}
+
+div.sourceCode,
+div.sourceCode pre.sourceCode {
+  color: #f8f8f2;
+}
+
+/* Normal */
+code span {
+  color: #f8f8f2;
+}
+
+/* Alert */
+code span.al {
+  color: #f07178;
+}
+
+/* Annotation */
+code span.an {
+  color: #d4d0ab;
+}
+
+/* Attribute */
+code span.at {
+  color: #00e0e0;
+}
+
+/* BaseN */
+code span.bn {
+  color: #d4d0ab;
+}
+
+/* BuiltIn */
+code span.bu {
+  color: #abe338;
+}
+
+/* ControlFlow */
+code span.cf {
+  font-weight: bold;
+  color: #ffa07a;
+}
+
+/* Char */
+code span.ch {
+  color: #abe338;
+}
+
+/* Constant */
+code span.cn {
+  color: #ffd700;
+}
+
+/* Comment */
+code span.co {
+  font-style: italic;
+  color: #f8f8f2;
+}
+
+/* CommentVar */
+code span.cv {
+  color: #ffd700;
+}
+
+/* Documentation */
+code span.do {
+  color: #f8f8f2;
+}
+
+/* DataType */
+code span.dt {
+  color: #ffa07a;
+}
+
+/* DecVal */
+code span.dv {
+  color: #d4d0ab;
+}
+
+/* Error */
+code span.er {
+  color: #f07178;
+  text-decoration: underline;
+}
+
+/* Extension */
+code span.ex {
+  font-weight: bold;
+  color: #00e0e0;
+}
+
+/* Float */
+code span.fl {
+  color: #d4d0ab;
+}
+
+/* Function */
+code span.fu {
+  color: #ffa07a;
+}
+
+/* Import */
+code span.im {
+  color: #abe338;
+}
+
+/* Information */
+code span.in {
+  color: #d4d0ab;
+}
+
+/* Keyword */
+code span.kw {
+  font-weight: bold;
+  color: #ffa07a;
+}
+
+/* Operator */
+code span.op {
+  color: #ffa07a;
+}
+
+/* Other */
+code span.ot {
+  color: #00e0e0;
+}
+
+/* Preprocessor */
+code span.pp {
+  color: #dcc6e0;
+}
+
+/* RegionMarker */
+code span.re {
+  background-color: #f8f8f2;
+  color: #00e0e0;
+}
+
+/* SpecialChar */
+code span.sc {
+  color: #abe338;
+}
+
+/* SpecialString */
+code span.ss {
+  color: #abe338;
+}
+
+/* String */
+code span.st {
+  color: #abe338;
+}
+
+/* Variable */
+code span.va {
+  color: #00e0e0;
+}
+
+/* VerbatimString */
+code span.vs {
+  color: #abe338;
+}
+
+/* Warning */
+code span.wa {
+  color: #dcc6e0;
+}
+
+.prevent-inlining {
+  content: "</";
+}
+
+/*# sourceMappingURL=237138eac573633834fecf06dae85922.css.map */
